Add support for server-browser communication through websockets
