+++ src/Kernel/ready.js
@@ -1,27 +1,12 @@
-// import { ipcRenderer } from 'electron'
 import { push } from 'react-router-redux'
+
+import actionClient from '../Kernel/actionClient'
 
 import * as Web3 from '../Actions/Web3'
 import * as Core from '../Actions/Core'
 import * as Accounts from '../Actions/Accounts'
 import * as Logs from '../Actions/Logs'
 import * as Settings from '../Actions/Settings'
-
-const initialSettings = {
-  googleAnalyticsTracking: false,
-  cpuAndMemoryProfiling: false,
-  verboseLogging: false,
-  firstRun: false,
-  server: {
-    hostname: "127.0.0.1",
-    port: 7545,
-    network_id: 5777,
-    total_accounts: 10,
-    unlocked_accounts: [],
-    mnemonic: "candy maple cake sugar pudding cream honey rich smooth crumble sweet treat",
-    vmErrorsOnRPCResponse: true
-  }
-}
 
 // This will be called before the very first render, so you can do whatever
 // you want here. The Redux Store is available at this point, so you can
@@ -31,8 +16,7 @@ export default function (store) {
   store.dispatch(Core.showTitleScreen())
 
   // Wait for the server to start...
-  //ipcRenderer.on(Core.SET_SERVER_STARTED, (sender, currentSettings) => {
-    const currentSettings = initialSettings
+  actionClient.on(Core.SET_SERVER_STARTED, (event, currentSettings) => {
     // Get current settings into the store
     store.dispatch(Settings.setSettings(currentSettings))
 
@@ -46,25 +30,25 @@ export default function (store) {
     store.dispatch(Core.getBlockSubscription())
 
     store.dispatch(Core.setServerStarted())
-  //})
-/*
+  })
+
   // Block polling happens in the chain process, and is passed through
   // the main process to the render process when there's a new block.
-  ipcRenderer.on(Core.SET_BLOCK_NUMBER, (event, number) => {
+  actionClient.on(Core.SET_BLOCK_NUMBER, (event, number) => {
     store.dispatch(Core.setBlockNumber(number))
   })
 
-  ipcRenderer.on(Core.SET_SYSTEM_ERROR, (event, error) => {
+  actionClient.on(Core.SET_SYSTEM_ERROR, (event, error) => {
     store.dispatch(Core.setSystemError(error))
   })
 
   // The server will send a second message that sets the mnemonic and hdpath
-  ipcRenderer.on(Core.SET_KEY_DATA, (event, data) => {
+  actionClient.on(Core.SET_KEY_DATA, (event, data) => {
     store.dispatch(Core.setKeyData(data.mnemonic, data.hdPath, data.privateKeys))
   })
 
-  ipcRenderer.on(Logs.ADD_LOG_LINES, (event, lines) => {
+  actionClient.on(Logs.ADD_LOG_LINES, (event, lines) => {
     store.dispatch(Logs.addLogLines(lines))
   })
-  */
+
 }
