+++ src/init.js
@@ -2,4 +2,4 @@
 var path = require("path");
 var appRoot = path.join(__dirname, '..');
 
-require('electron-compile').init(appRoot, require.resolve('./main'));
+require('electron-compile').init(appRoot, require.resolve('./main-electron'));
